refactor: move Flex component from experimental to dedicated folder
